pyproject.toml
[tool.poetry]
name = "human-anomaly-behavior-recognition"
version = "0.1.0"
description = ""
authors = ["mcdnmd <kpoltoradnev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
opencv-python = "^4.7.0.72"
pandas = "^1.5.3"
torch = "^2.0.0"
loguru = "^0.6.0"
torchvision = "^0.15.1"
gradio = "^3.24.1"
matplotlib = "3.2.2"
pillow = "^9.5.0"
torch-geometric = "2.3.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
